"FIx" Active System Administrator Users' Email addresses
Addresses the issue of email addresses in refreshed and created sandboxes where the email address of all Users is altered.  This class can be employed in Production when performing these activities to ensure that system administrator users can access the sandbox without having to engage the user who initiated the action.